refactor(gateway,session): split oversized runtime modules
Issue: #1174

- decompose gateway_openresponses into auth, websocket, multi-channel status, webchat, and dedicated tests modules

- decompose tau-session lib into integrity, storage, locking, and dedicated tests modules

- preserve API/runtime behavior while reducing file size and clarifying ownership with module docs
